Add files via upload
added spheres, rays, and the intersections that are found when rays are shot out. currently able to draw 2d sphere and transform the shape of the sphere into different shapes. shearing, scaling, translating, and rotating. 

Descriptions of how to use these functions will be added as comments. will likely add functions that do most of the function calls for ease of use (will likely need templating for multiple use cases)